Refactor `views.document` to avoid circular dependencies
