+++ c2corg_api/views/area.py
@@ -1,14 +1,15 @@
 from c2corg_api.models.area import schema_area, Area, schema_update_area, \
-    schema_listing_area, AREA_TYPE
+    AREA_TYPE
 from c2corg_api.models.area_association import update_area
 from c2corg_api.models.cache_version import update_cache_version_for_area
 from c2corg_api.models.document import UpdateType
 from c2corg_api.views.document_info import DocumentInfoRest
+from c2corg_api.views.document_schemas import area_documents_config
 from c2corg_common.fields_area import fields_area
 from cornice.resource import resource, view
 
 from c2corg_api.views.document import DocumentRest, make_validator_create, \
-    make_validator_update, GetDocumentsConfig
+    make_validator_update
 from c2corg_api.views import cors_policy, restricted_json_view
 from c2corg_api.views.validation import validate_id, validate_pagination, \
     validate_lang_param, validate_preferred_lang_param, validate_lang
@@ -48,9 +49,6 @@ class AreaRest(DocumentRest):
                 raise HTTPBadRequest('No permission to change the geometry')
 
         return self._put(Area, schema_area, after_update=update_associations)
-
-area_documents_config = GetDocumentsConfig(
-    AREA_TYPE, Area, schema_listing_area, include_areas=False)
 
 
 @resource(path='/areas/{id}/{lang}/info', cors_policy=cors_policy)
