+++ c2corg_api/views/user_profile.py
@@ -1,10 +1,11 @@
 
 from c2corg_api.models.user_profile import schema_update_user_profile, \
     UserProfile, schema_user_profile, schema_internal_user_profile, \
-    schema_listing_user_profile, USERPROFILE_TYPE
+    USERPROFILE_TYPE
+from c2corg_api.views.document_schemas import user_profile_documents_config
 from cornice.resource import resource
 
-from c2corg_api.views.document import DocumentRest, GetDocumentsConfig
+from c2corg_api.views.document import DocumentRest
 from c2corg_api.views import cors_policy, restricted_json_view, restricted_view
 from c2corg_api.views.validation import validate_id, validate_pagination, \
     validate_lang_param, validate_preferred_lang_param
@@ -50,7 +51,3 @@ class UserProfileRest(DocumentRest):
         if locales:
             for locale in locales:
                 locale['title'] = ''
-
-
-user_profile_documents_config = GetDocumentsConfig(
-    USERPROFILE_TYPE, UserProfile, schema_listing_user_profile)
